Models/Gen_Ai/requirements.txt
pandas
chromadb
google-generativeai
fastapi
uvicorn
streamlit
python-dotenv